+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/domain/commons/CommonFilterService.java
@@ -22,8 +22,8 @@ package ch.puzzle.itc.mobiliar.business.domain.commons;
 
 import ch.puzzle.itc.mobiliar.business.deploy.boundary.DeploymentService;
 import ch.puzzle.itc.mobiliar.business.utils.JpaWildcardConverter;
-import ch.puzzle.itc.mobiliar.common.util.CustomFilter;
-import ch.puzzle.itc.mobiliar.common.util.CustomFilter.ComperatorFilterOption;
+import ch.puzzle.itc.mobiliar.business.deploy.entity.CustomFilter;
+import ch.puzzle.itc.mobiliar.business.deploy.entity.CustomFilter.ComperatorFilterOption;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
