+++ AMW_web/src/main/java/ch/puzzle/itc/mobiliar/presentation/deploy/DeployScreenController.java
@@ -33,7 +33,7 @@ import ch.puzzle.itc.mobiliar.business.resourcegroup.boundary.ResourceGroupLocat
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceEntity;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceGroupEntity;
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceTypeEntity;
-import ch.puzzle.itc.mobiliar.common.util.CustomFilter;
+import ch.puzzle.itc.mobiliar.business.deploy.entity.CustomFilter;
 import ch.puzzle.itc.mobiliar.common.util.DefaultResourceTypeDefinition;
 import ch.puzzle.itc.mobiliar.common.util.Tuple;
 import ch.puzzle.itc.mobiliar.presentation.deploy.DeployScreenDataProvider.deployscreenColDescriptor;
