+++ AMW_web/src/main/java/ch/puzzle/itc/mobiliar/presentation/shakedowntest/ShakedownTestScreenController.java
@@ -40,7 +40,7 @@ import ch.puzzle.itc.mobiliar.business.shakedown.entity.ShakedownTestEntity.shak
 import ch.puzzle.itc.mobiliar.business.shakedown.entity.ShakedownTestOrder;
 import ch.puzzle.itc.mobiliar.business.shakedown.event.ShakedownTestEvent;
 import ch.puzzle.itc.mobiliar.business.shakedown.event.ShakedownTestEvent.ShakedownTestEventType;
-import ch.puzzle.itc.mobiliar.common.util.CustomFilter;
+import ch.puzzle.itc.mobiliar.business.deploy.entity.CustomFilter;
 import ch.puzzle.itc.mobiliar.common.util.DefaultResourceTypeDefinition;
 import ch.puzzle.itc.mobiliar.common.util.Tuple;
 import ch.puzzle.itc.mobiliar.presentation.shakedowntest.ShakedownTestScreenDataProvider.shakedownTestScreenColDescriptor;
